ARROW-13803: [C++] Don't read past end of buffer in BitUtil::SetBitmap
I can only get this to actually fail on M1 ARM with optimizations, but what happens here is we read one past the end of a buffer. On M1, this actually ends up in an unmapped region, crashing the program.

On Linux/x64, I tried AddressSanitizer and Valgrind but neither caught the access, oddly enough.

Closes #11070 from lidavidm/arrow-13803

Lead-authored-by: Yibo Cai <yibo.cai@arm.com>
Co-authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>