ARROW-14157: [C++] Refactor Abseil to its own macro
We will need Abseil for the GCS C++ SDK (aka `google-cloud-cpp`), so
split it to its own macro. Use the `abseil_ep` target as a sentinel to
make the macro idempotent.

Abseil creates many (over 100) libraries, each relatively small. These
libraries depend on each other. Document how to extract this information
using a shell command-line for future updates. We will not need all
these libraries, but it is easier to describe all their dependencies
that pick and choose which ones are used.

Closes #11261 from coryan/ARROW-14157-refactor-abseil-dependency

Authored-by: Carlos O'Ryan <coryan@google.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>